api/pyproject.toml
[project]
name = "insidertrack-api"
version = "0.1.0"
description = "InsiderTrack API for Brazilian CVM insider trading data"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic-settings>=2.1.0",
    "psycopg2-binary>=2.9.9",
    "httpx>=0.26.0",
    "pdfplumber>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "pyright>=1.1.0",
]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "standard"
reportMissingImports = "warning"
